Fix signaure of mve_cb_Alloc

+++ d2x-rebirth/include/libmve.h
@@ -32,7 +32,7 @@ typedef unsigned int (*mve_cb_Read)(void *stream,
                                     void *buffer,
                                     unsigned int count);
 
-typedef void *(*mve_cb_Alloc)(unsigned int size);
+typedef void *(*mve_cb_Alloc)(size_t size);
 typedef void (*mve_cb_Free)(void *ptr);
 
 typedef void (*mve_cb_ShowFrame)(unsigned char *buffer, int dstx, int dsty, int bufw, int bufh, int sw, int sh);
